Mock 객체 Stubbing - 리턴값 있는 메소드의 리턴값 조작하기, void 메소드의 예외 조작하기

+++ src/main/java/me/whiteship/inflearnthejavatest/study/StudyRepository.java
@@ -3,5 +3,5 @@ package me.whiteship.inflearnthejavatest.study;
 import me.whiteship.inflearnthejavatest.domain.Study;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudyRepository extends JpaRepository<Study, Integer> {
+public interface StudyRepository extends JpaRepository<Study, Long> {
 }
